Adopt SF-0023 ProgressManager
Progress now flows down into SwiftNASR as a `Subprogress` minted from the
caller's own `ProgressManager`, rather than up out of it as a `Progress`
handed to a handler block. Six public signatures trade their
`withProgress:` handler for `progress: consuming Subprogress? = nil`:
`NASR.load`, `NASR.parse`, `Loader.load`, `Downloader.load`,
`Distribution.readFile`, and `Distribution.readFileRaw`.

The platform floor rises to macOS/iOS/tvOS/watchOS/visionOS 27 and the
manifest to tools-version 6.4, which is where `ProgressManager` starts
existing: it carries no lower availability annotation on Apple platforms,
it is `@available(FoundationPreview 6.4)` on Linux, and `MacOSVersion.v27`
is itself `@available(_PackageDescription 6.4)`. CI follows: the Darwin
legs collapse into one `xcode-27` job driven by `maxim-lobanov/setup-xcode`
(no earlier image can supply the macOS 27 SDK), the Linux 6.3 leg is
dropped, and `.github/actionlint.yaml` declares `xcode-27` so
`actionlint` keeps checking runner labels.

`DirectoryDistribution` double-counted every file it read, adding each
line's length on top of the length of the chunk the line arrived in and
reporting 35 bytes read from a 21-byte file. `Progress` clamped
`fractionCompleted` to 1.0 and hid it; `ProgressManager` reports the
overshoot. The file's byte accounting is now pinned through the new
`completedByteCount`/`totalByteCount` metadata, which every read and
download reports alongside its unit counts.

Also: `CSVParser` sheds its vestigial `progress` and `bytesRead`
requirements, a diagnosed record no longer stalls the fixed-width parse
progress short of its total, and the E2E progress bar follows
`Observations.untilFinished` over the observable root manager instead of
re-reading `fractionCompleted` on a 10 Hz timer.

+++ Tests/SwiftNASR_E2E/RecordTypeRegistry.swift
@@ -4,13 +4,13 @@ import SwiftNASR
 struct RecordTypeInfo {
   let recordType: RecordType
   let displayName: String
-  let txtWeight: Int64?  // nil if not available in TXT
-  let csvWeight: Int64?  // nil if not available in CSV
+  let txtWeight: Int?  // nil if not available in TXT
+  let csvWeight: Int?  // nil if not available in CSV
 
   var availableInTXT: Bool { txtWeight != nil }
   var availableInCSV: Bool { csvWeight != nil }
 
-  func weight(isCSV: Bool) -> Int64 {
+  func weight(isCSV: Bool) -> Int {
     (isCSV ? csvWeight : txtWeight) ?? 1
   }
 }
